+++ packages/data/rpc_data/lib/src/repository/i_data_repository.dart
@@ -11,6 +11,8 @@ abstract interface class IDataRepository {
 
   Future<DataRecord?> get(GetRecordRequest request);
 
+  Future<List<DataRecord>> getMany(GetRecordsRequest request);
+
   Future<ListRecordsResponse> list(ListRecordsRequest request);
 
   Future<DataRecord> update(UpdateRecordRequest request);